requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-yasg==1.21.7
psycopg2-binary==2.9.9
redis==5.2.1
